+++ src/lib.rs
@@ -11,3 +11,5 @@
 pub mod error;
 pub mod r0;
 pub mod unversioned;
+
+mod serde;