+++ src/r0/sync/sync_events.rs
@@ -1,6 +1,6 @@
 //! [GET /_matrix/client/r0/sync](https://matrix.org/docs/spec/client_server/r0.4.0.html#get-matrix-client-r0-sync)
 
-use std::collections::HashMap;
+use std::{collections::HashMap, time::Duration};
 
 use js_int::UInt;
 use ruma_api::{ruma_api, Outgoing};
@@ -48,8 +48,9 @@ ruma_api! {
         pub set_presence: Option<SetPresence>,
         /// The maximum time to poll in milliseconds before returning this request.
         #[serde(skip_serializing_if = "Option::is_none")]
+        #[serde(with = "crate::serde::duration::opt_ms")]
         #[ruma_api(query)]
-        pub timeout: Option<UInt>,
+        pub timeout: Option<Duration>,
     }
 
     response {