+++ src/r0/typing/create_typing_event.rs
@@ -1,6 +1,7 @@
 //! [PUT /_matrix/client/r0/rooms/{roomId}/typing/{userId}](https://matrix.org/docs/spec/client_server/r0.4.0.html#put-matrix-client-r0-rooms-roomid-typing-userid)
 
-use js_int::UInt;
+use std::time::Duration;
+
 use ruma_api::ruma_api;
 use ruma_identifiers::{RoomId, UserId};
 
@@ -20,7 +21,8 @@ ruma_api! {
         pub room_id: RoomId,
         /// The length of time in milliseconds to mark this user as typing.
         #[serde(skip_serializing_if = "Option::is_none")]
-        pub timeout: Option<UInt>,
+        #[serde(default, with = "crate::serde::duration::opt_ms")]
+        pub timeout: Option<Duration>,
         /// Whether the user is typing or not. If `false`, the `timeout` key can be omitted.
         pub typing: bool,
         /// The user who has started to type.