+++ src/serde.rs
@@ -1,0 +1,3 @@
+//! Modules to hold functions for de-/serializing remote types
+
+pub mod duration;